Fixed OK status and pagination code checks
The pagination code check used to append the code
even if there was no string contents in the variable
which is not exactly wrong but also not very useful,
this has been fixed now and condensed into one conditional
statement.

The OK Status used to only check for 200 but HTTP defines
all statuses [200,300[ as OK-status. The check was ammended to
accomodate for this.